Moved Firewall classes to Pipeline namespace
