+++ tests/Pipeline/AllowedServerFirewallStepTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yumerov\MaxiBot\Firewalls;
+namespace Yumerov\MaxiBot\Pipeline;
 
 use Discord\Discord;
 use PHPUnit\Framework\MockObject\Exception;
@@ -10,7 +10,7 @@ use Yumerov\MaxiBot\DTO\EnvDTO;
 use Yumerov\MaxiBot\Mocks\Message;
 use Yumerov\MaxiBot\Mocks\Thread;
 
-class AllowedServerFirewallTest extends TestCase
+class AllowedServerFirewallStepTest extends TestCase
 {
 
     private Message $message;
@@ -76,9 +76,9 @@ class AllowedServerFirewallTest extends TestCase
     /**
      * @throws Exception
      */
-    private function initFirewall(string $allowedServers): AllowedServerFirewall
+    private function initFirewall(string $allowedServers): AllowedServerFirewallStep
     {
-        return new AllowedServerFirewall(
+        return new AllowedServerFirewallStep(
             $this->discord,
             $this->message,
             $this->createMock(LoggerInterface::class),
